Add feature to check if account is added/removed
Why:

CT-25

This change addresses the need by:

We utilize the BeautifulSoup Python library to read and parse a given
bank's webpage for HTML elements. We then extract a subset of the HTML
and count the number of elements related to the bank accounts. If the
count matches what we expect to see on the page, then no accounts have
been added/removed. If it doesn't match, we indicate as such and edit
the config file to reflect the difference.